Rewrote the FontError to be unified enum for all fonts
